requirements.txt
pyglet
flake8
nose
coverage
msgpack-python